+++ packages/data-schema/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @aws-amplify/data-schema
 
+## 1.5.0
+
+### Minor Changes
+
+- 6028bf7: Add AI routes to client
+
+## 1.4.1
+
+### Patch Changes
+
+- ed16182: adds selectionSet types to mutation operations
+
 ## 1.4.0
 
 ### Minor Changes